incorporated session timer in the UI

+++ lib/f1_types/packet_1_session_data.py
@@ -595,9 +595,9 @@ class PacketSessionData:
         self.m_sessionType: int
         self.m_trackId: TrackID
         self.m_formula: int
-        self.m_sessionTimeLeft: float
-        self.m_sessionDuration: float
-        self.m_pitSpeedLimit: float
+        self.m_sessionTimeLeft: int
+        self.m_sessionDuration: int
+        self.m_pitSpeedLimit: int
         self.m_gamePaused: bool
         self.m_isSpectating: bool
         self.m_spectatorCarIndex: int
